Changing from code-only to real-world example
Since separating the Tank studio install from the project area is now
built-in functionality as of Tank v0.13, the example configuration for
working around the lack of that feature is no longer needed.  But, we
still need an example of how to use multiple roots for actual data, so
this configuration has been changed to reflect that.